Status effects on visitors
- Fixing some issue with task targets
- Creation of effects on task execution
- Storing effects on DB
- Provide effects readable in GraphQL interface
- Consider effects during fight evaluations

Missing: Read and display in frontend

+++ migrations/2019-07-09-161445_create_buildings/up.sql
@@ -5,7 +5,7 @@ CREATE TABLE buildings (
     y INT NOT NULL,
     building_type BUILDING_TYPE NOT NULL,
     building_range REAL,
-    attack_power REAL,
+    attack_power INT,
     attacks_per_cycle INT,
     creation TIMESTAMP NOT NULL
 )
